!R (MPFR) Explicit Memset(0) work-arounds and memcpy() compatibility
Don't depend on auto-initialization by MPFloat if it detects that it's been memset before assignment/etc.
Require explicit checks, to avoid hiding this issue and reduce overhead for this.
Tweaked for faster set of CTimeValue arrays, by pre-allocating 1 'zeroed out' type and copying it over.

Now memcpy() compatible with custom MPFR edits to allow POD-style.
Updated bugged network serialization (and other serializations)
Was getting size of CTimeValue but serializing as str, which could have a larger size == random malformed data.

Plus CTimeValue tabs/etc. fixed.
And PhysInterface uses NaN instead of memHack().

+++ Code/CryEngine/CrySystem/StreamEngine/StreamEngine.cpp
@@ -1253,7 +1253,7 @@ void CStreamEngine::DrawStatistics()
 		         (float)stats.hddInfo.nActualReadBandwidth / (1024 * 1024), (float)stats.hddInfo.nAverageActualReadBandwidth / (1024 * 1024));
 		DrawText(tx, ty += ystep, clText, "\t  Seek: %1.2f GB - Active: %2.1f%%(%2.1f%%)",
 		         (float)stats.hddInfo.nAverageSeekOffset / (1024 * 1024),
-		         stats.hddInfo.fActiveDuringLastSecond, stats.hddInfo.fAverageActiveTime);
+		         (float)stats.hddInfo.fActiveDuringLastSecond, stats.hddInfo.fAverageActiveTime);
 	}
 	// Optical stats
 	if (stats.discInfo.nTotalRequestCount > 0)
@@ -1265,7 +1265,7 @@ void CStreamEngine::DrawStatistics()
 		         (float)stats.discInfo.nActualReadBandwidth / (1024 * 1024), (float)stats.discInfo.nAverageActualReadBandwidth / (1024 * 1024));
 		DrawText(tx, ty += ystep, clText, "\t  Seek: %1.2f GB - Active: %2.1f%%(%2.1f%%)",
 		         (float)stats.discInfo.nAverageSeekOffset / (1024 * 1024),
-		         stats.discInfo.fActiveDuringLastSecond, stats.discInfo.fAverageActiveTime);
+		         (float)stats.discInfo.fActiveDuringLastSecond, stats.discInfo.fAverageActiveTime);
 	}
 	DrawText(tx, ty += ystep, clText, "Mem : Request: %3d|%5d (%4d MB)",
 	         stats.memoryInfo.nRequestCount, stats.memoryInfo.nTotalRequestCount, (stats.memoryInfo.nTotalBytesRead / (1024 * 1024)));
